all tests pass slightly cleaner
