+++ rpc/events.go
@@ -1,16 +1,9 @@
 package rpc
 
 import (
-	"context"
-	"encoding/json"
-
 	"github.com/NethermindEth/juno/blockchain"
-	"github.com/NethermindEth/juno/core"
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/juno/feed"
 	"github.com/NethermindEth/juno/jsonrpc"
-	"github.com/NethermindEth/juno/sync"
-	"github.com/sourcegraph/conc"
 )
 
 type EventsArg struct {
@@ -55,335 +48,6 @@ type SubscriptionID struct {
 /****************************************************
 		Events Handlers
 *****************************************************/
-
-func (h *Handler) SubscribeNewHeads(ctx context.Context, blockID *BlockID) (*SubscriptionID, *jsonrpc.Error) {
-	w, ok := jsonrpc.ConnFromContext(ctx)
-	if !ok {
-		return nil, jsonrpc.Err(jsonrpc.MethodNotFound, nil)
-	}
-
-	startHeader, latestHeader, rpcErr := h.getStartAndLatestHeaders(blockID)
-	if rpcErr != nil {
-		return nil, rpcErr
-	}
-
-	id := h.idgen()
-	subscriptionCtx, subscriptionCtxCancel := context.WithCancel(ctx)
-	sub := &subscription{
-		cancel: subscriptionCtxCancel,
-		conn:   w,
-	}
-	h.mu.Lock()
-	h.subscriptions[id] = sub
-	h.mu.Unlock()
-
-	headerSub := h.newHeads.Subscribe()
-	reorgSub := h.reorgs.Subscribe() // as per the spec, reorgs are also sent in the new heads subscription
-	sub.wg.Go(func() {
-		defer func() {
-			h.unsubscribe(sub, id)
-			headerSub.Unsubscribe()
-			reorgSub.Unsubscribe()
-		}()
-
-		var wg conc.WaitGroup
-
-		wg.Go(func() {
-			if err := h.sendHistoricalHeaders(subscriptionCtx, startHeader, latestHeader, w, id); err != nil {
-				h.log.Errorw("Error sending old headers", "err", err)
-				return
-			}
-		})
-
-		wg.Go(func() {
-			h.processNewHeaders(subscriptionCtx, headerSub, w, id)
-		})
-
-		wg.Go(func() {
-			h.processReorgs(subscriptionCtx, reorgSub, w, id)
-		})
-
-		wg.Wait()
-	})
-
-	return &SubscriptionID{ID: id}, nil
-}
-
-func (h *Handler) SubscribePendingTxs(ctx context.Context, getDetails *bool, senderAddr []felt.Felt) (*SubscriptionID, *jsonrpc.Error) {
-	w, ok := jsonrpc.ConnFromContext(ctx)
-	if !ok {
-		return nil, jsonrpc.Err(jsonrpc.MethodNotFound, nil)
-	}
-
-	if len(senderAddr) > maxEventFilterKeys {
-		return nil, ErrTooManyAddressesInFilter
-	}
-
-	id := h.idgen()
-	subscriptionCtx, subscriptionCtxCancel := context.WithCancel(ctx)
-	sub := &subscription{
-		cancel: subscriptionCtxCancel,
-		conn:   w,
-	}
-	h.mu.Lock()
-	h.subscriptions[id] = sub
-	h.mu.Unlock()
-
-	pendingTxsSub := h.pendingTxs.Subscribe()
-	sub.wg.Go(func() {
-		defer func() {
-			h.unsubscribe(sub, id)
-			pendingTxsSub.Unsubscribe()
-		}()
-
-		h.processPendingTxs(subscriptionCtx, getDetails != nil && *getDetails, senderAddr, pendingTxsSub, w, id)
-	})
-
-	return &SubscriptionID{ID: id}, nil
-}
-
-func (h *Handler) processPendingTxs(
-	ctx context.Context,
-	getDetails bool,
-	senderAddr []felt.Felt,
-	pendingTxsSub *feed.Subscription[[]core.Transaction],
-	w jsonrpc.Conn,
-	id uint64,
-) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case pendingTxs := <-pendingTxsSub.Recv():
-			filteredTxs := h.filterTxs(pendingTxs, getDetails, senderAddr)
-			if err := h.sendPendingTxs(w, filteredTxs, id); err != nil {
-				h.log.Warnw("Error sending pending transactions", "err", err)
-				return
-			}
-		}
-	}
-}
-
-func (h *Handler) filterTxs(pendingTxs []core.Transaction, getDetails bool, senderAddr []felt.Felt) interface{} {
-	if getDetails {
-		return h.filterTxDetails(pendingTxs, senderAddr)
-	}
-	return h.filterTxHashes(pendingTxs, senderAddr)
-}
-
-func (h *Handler) filterTxDetails(pendingTxs []core.Transaction, senderAddr []felt.Felt) []*Transaction {
-	filteredTxs := make([]*Transaction, 0, len(pendingTxs))
-	for _, txn := range pendingTxs {
-		if h.shouldIncludeTx(txn, senderAddr) {
-			filteredTxs = append(filteredTxs, AdaptTransaction(txn))
-		}
-	}
-	return filteredTxs
-}
-
-func (h *Handler) filterTxHashes(pendingTxs []core.Transaction, senderAddr []felt.Felt) []felt.Felt {
-	filteredTxHashes := make([]felt.Felt, 0, len(pendingTxs))
-	for _, txn := range pendingTxs {
-		if h.shouldIncludeTx(txn, senderAddr) {
-			filteredTxHashes = append(filteredTxHashes, *txn.Hash())
-		}
-	}
-	return filteredTxHashes
-}
-
-func (h *Handler) shouldIncludeTx(txn core.Transaction, senderAddr []felt.Felt) bool {
-	if len(senderAddr) == 0 {
-		return true
-	}
-
-	//
-	switch t := txn.(type) {
-	case *core.InvokeTransaction:
-		for _, addr := range senderAddr {
-			if t.SenderAddress.Equal(&addr) {
-				return true
-			}
-		}
-	case *core.DeclareTransaction:
-		for _, addr := range senderAddr {
-			if t.SenderAddress.Equal(&addr) {
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
-func (h *Handler) sendPendingTxs(w jsonrpc.Conn, result interface{}, id uint64) error {
-	resp, err := json.Marshal(SubscriptionResponse{
-		Version: "2.0",
-		Method:  "starknet_subscriptionPendingTransactions",
-		Params: map[string]interface{}{
-			"subscription_id": id,
-			"result":          result,
-		},
-	})
-	if err != nil {
-		return err
-	}
-
-	_, err = w.Write(resp)
-	return err
-}
-
-// getStartAndLatestHeaders gets the start and latest header for the subscription
-func (h *Handler) getStartAndLatestHeaders(blockID *BlockID) (*core.Header, *core.Header, *jsonrpc.Error) {
-	if blockID == nil || blockID.Latest {
-		return nil, nil, nil
-	}
-
-	if blockID.Pending {
-		return nil, nil, ErrCallOnPending
-	}
-
-	latestHeader, err := h.bcReader.HeadsHeader()
-	if err != nil {
-		return nil, nil, ErrInternal
-	}
-
-	startHeader, rpcErr := h.blockHeaderByID(blockID)
-	if rpcErr != nil {
-		return nil, nil, rpcErr
-	}
-
-	// TODO(weiihann): also, reuse this function in other places
-	if latestHeader.Number >= maxBlocksBack && startHeader.Number <= latestHeader.Number-maxBlocksBack {
-		return nil, nil, ErrTooManyBlocksBack
-	}
-
-	return startHeader, latestHeader, nil
-}
-
-// sendHistoricalHeaders sends a range of headers from the start header until the latest header
-func (h *Handler) sendHistoricalHeaders(
-	ctx context.Context,
-	startHeader *core.Header,
-	latestHeader *core.Header,
-	w jsonrpc.Conn,
-	id uint64,
-) error {
-	if startHeader == nil {
-		return nil
-	}
-
-	var err error
-
-	lastHeader := startHeader
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-			if err := h.sendHeader(w, lastHeader, id); err != nil {
-				return err
-			}
-
-			if lastHeader.Number == latestHeader.Number {
-				return nil
-			}
-
-			lastHeader, err = h.bcReader.BlockHeaderByNumber(lastHeader.Number + 1)
-			if err != nil {
-				return err
-			}
-		}
-	}
-}
-
-func (h *Handler) bufferNewHeaders(ctx context.Context, headerSub *feed.Subscription[*core.Header], newHeadersChan chan<- *core.Header) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case header := <-headerSub.Recv():
-			newHeadersChan <- header
-		}
-	}
-}
-
-func (h *Handler) processNewHeaders(ctx context.Context, headerSub *feed.Subscription[*core.Header], w jsonrpc.Conn, id uint64) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case header := <-headerSub.Recv():
-			if err := h.sendHeader(w, header, id); err != nil {
-				h.log.Warnw("Error sending header", "err", err)
-				return
-			}
-		}
-	}
-}
-
-// sendHeader creates a request and sends it to the client
-func (h *Handler) sendHeader(w jsonrpc.Conn, header *core.Header, id uint64) error {
-	resp, err := json.Marshal(SubscriptionResponse{
-		Version: "2.0",
-		Method:  "starknet_subscriptionNewHeads",
-		Params: map[string]any{
-			"subscription_id": id,
-			"result":          adaptBlockHeader(header),
-		},
-	})
-	if err != nil {
-		return err
-	}
-	_, err = w.Write(resp)
-	return err
-}
-
-func (h *Handler) processReorgs(ctx context.Context, reorgSub *feed.Subscription[*sync.ReorgData], w jsonrpc.Conn, id uint64) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case reorg := <-reorgSub.Recv():
-			if err := h.sendReorg(w, reorg, id); err != nil {
-				h.log.Warnw("Error sending reorg", "err", err)
-				return
-			}
-		}
-	}
-}
-
-func (h *Handler) sendReorg(w jsonrpc.Conn, reorg *sync.ReorgData, id uint64) error {
-	resp, err := json.Marshal(jsonrpc.Request{
-		Version: "2.0",
-		Method:  "starknet_subscriptionReorg",
-		Params: map[string]any{
-			"subscription_id": id,
-			"result":          reorg,
-		},
-	})
-	if err != nil {
-		return err
-	}
-	_, err = w.Write(resp)
-	return err
-}
-
-func (h *Handler) Unsubscribe(ctx context.Context, id uint64) (bool, *jsonrpc.Error) {
-	w, ok := jsonrpc.ConnFromContext(ctx)
-	if !ok {
-		return false, jsonrpc.Err(jsonrpc.MethodNotFound, nil)
-	}
-	h.mu.Lock()
-	sub, ok := h.subscriptions[id]
-	h.mu.Unlock() // Don't defer since h.unsubscribe acquires the lock.
-	if !ok || !sub.conn.Equal(w) {
-		return false, ErrSubscriptionNotFound
-	}
-	sub.cancel()
-	sub.wg.Wait() // Let the subscription finish before responding.
-	return true, nil
-}
-
 // Events gets the events matching a filter
 //
 // It follows the specification defined here:
